Move files in com.google.alpollo to the separate packages for logic
