+++ src/main/java/com/google/alpollo/helpers/AnalysisHelper.java
@@ -1,4 +1,4 @@
-package com.google.alpollo;
+package com.google.alpollo.helpers;
 
 import java.io.IOException;
 import java.util.ArrayList;
