+++ src/main/java/com/google/alpollo/servlets/AnalysisInfoServlet.java
@@ -1,10 +1,12 @@
-package com.google.alpollo;
+package com.google.alpollo.servlets;
 
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import com.google.alpollo.database.SongDataBase;
 import com.google.alpollo.model.AnalysisInfo;
 import com.google.gson.Gson;
 import com.google.gson.JsonIOException;
