+++ src/main/java/com/google/alpollo/servlets/EntityServlet.java
@@ -1,4 +1,4 @@
-package com.google.alpollo;
+package com.google.alpollo.servlets;
 
 import java.io.IOException;
 import java.util.List;
@@ -6,6 +6,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import com.google.alpollo.helpers.AnalysisHelper;
+import com.google.alpollo.helpers.ConfigHelper;
 import com.google.alpollo.model.Lyrics;
 import com.google.alpollo.model.SongEntity;
 import com.google.cloud.language.v1.Entity;
