+++ src/test/java/com/google/alpollo/AnalysisTest.java
@@ -1,7 +1,10 @@
 package com.google.alpollo;
 
+import com.google.alpollo.helpers.AnalysisHelper;
 import com.google.alpollo.model.Lyrics;
 import com.google.alpollo.model.SongSentiment;
+import com.google.alpollo.servlets.EntityServlet;
+import com.google.alpollo.servlets.SentimentServlet;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Before;
